generate_token.py
import os.path
from google.auth.transport.requests import Request
from google.oauth2.credentials import Credentials
from google_auth_oauthlib.flow import InstalledAppFlow

# If you change these scopes, delete the file token.json.
SCOPES = ['https://www.googleapis.com/auth/calendar.readonly']

def main():
    creds = None
    # The file token.json stores the user's access and refresh tokens.
    if os.path.exists('config/token.json'):
        creds = Credentials.from_authorized_user_file('token.json', SCOPES)

    # If there are no (valid) credentials available, let the user log in.
    if not creds or not creds.valid:
        if creds and creds.expired and creds.refresh_token:
            creds.refresh(Request())
        else:
            flow = InstalledAppFlow.from_client_secrets_file(
                'credentials.json', SCOPES)
            # This opens your local browser
            creds = flow.run_local_server(port=0)

        # Save the credentials for the next run
        with open('config/token.json', 'w') as token:
            token.write(creds.to_json())
            print("✅ Success! token.json has been created.")

if __name__ == '__main__':
    main()